push down extra preicates trough join
